Move video-preview hoc to directory
